PvPLogic updated

+++ KipoBot/Game/Abilities/Punch.cs
@@ -22,7 +22,7 @@ namespace KipoBot.Game.Abilities
         public override Task Use(SocketCommandContext ctx, Pet attacker, Pet target)
         {
             int damage = Damage(attacker);
-            target.health -= (byte)damage;
+            target.health -= (short)damage;
             ctx.Channel.SendMessageAsync($"{attacker.name} attacked {target.name} using {name} ability!\n" +
                 $"{target.name} Health - {target.health}");
